+++ src/main/java/ru/fds/tavrzcms3/service/PledgeAgreementService.java
@@ -14,6 +14,7 @@ import ru.fds.tavrzcms3.specification.impl.SpecificationBuilderImpl;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -221,7 +222,7 @@ public class PledgeAgreementService {
 
         for(Field field : PledgeAgreement.class.getDeclaredFields()){
             if(searchParam.containsKey(field.getName())){
-                if((field.getType() == String.class || field.getType() == double.class) && !searchParam.get(field.getName()).isEmpty()){
+                if((field.getType() == String.class || field.getType() == BigDecimal.class) && !searchParam.get(field.getName()).isEmpty()){
                     SearchCriteria searchCriteria = SearchCriteria.builder()
                             .key(field.getName())
                             .value(searchParam.get(field.getName()))
